Auto generate model key
