+++ src/handler/save.ts
@@ -5,4 +5,4 @@ import { __conn } from '../connections/connection-handler';
  */
 export const save = (key: string, data, collection?): Promise<any> => {
   return (collection ? collection : __conn.getCollection()).upsert(key, data);
-}
+};
