+++ src/model/model-base.ts
@@ -1,11 +1,11 @@
-import {validateSchema} from "..";
-import { save } from "../handler/save";
+import { validateSchema } from '..';
+import { save } from '../handler/save';
 
 export class ModelBase {
-    constructor(public __data, public __collection, public __schema) {}
+  constructor(public __data, public __collection, public __schema) {}
 
-    save(key) {
-        validateSchema(this.__data, this.__schema);
-        return save(key, this.__data, this.__collection);
-    }
+  save(key) {
+    validateSchema(this.__data, this.__schema);
+    return save(key, this.__data, this.__collection);
+  }
 }
