+++ src/utils/generate-uuid.ts
@@ -1,0 +1,5 @@
+import { v4 as uuidv4 } from 'uuid';
+
+export const generateUUID = () => {
+  return uuidv4();
+};
